Move Converter to base class and add wrapper
